+++ package.json
@@ -61,7 +61,7 @@
     "react-dom": "^16.4.0",
     "react-modal": "^3.4.5",
     "react-redux": "^5.0.7",
-    "react-toggle": "^4.0.2",
+    "react-router-dom": "^4.3.1",
     "redux": "^4.0.0",
     "redux-thunk": "^2.3.0",
     "shave": "^2.1.3",
