+++ src/components/ListArticles/_ListArticles.scss
@@ -1,3 +1,7 @@
+.article-create-link {
+  margin-bottom: $indent-xl;
+  text-align: right;
+}
 .articles-list {
 	&__item {
 		padding: $indent-l;
