--- src/components/ListArticles/components/AddNewArticle/components/Form/index.js
@@ -1,3 +1,0 @@
-import AddNewForm from './Form';
-
-export default AddNewForm;
