--- src/components/ListArticles/components/AddNewArticle/index.js
@@ -1,3 +1,0 @@
-import AddNewArticle from './AddNewArticle';
-
-export default AddNewArticle;
