+++ src/components/Toolbar/index.js
@@ -1,0 +1,3 @@
+import UserLinks from './container/Toolbar';
+
+export default UserLinks;
