--- src/components/User/_User.scss
@@ -1,24 +1,0 @@
-.tabs{
-  display: flex;
-  align-items: baseline;
-  margin-bottom: 50px;
-  &__item{
-    &:not(:last-child) {
-      margin-right: $indent-xl;
-    }
-    &.is-active{
-      text-decoration: underline;
-    }
-  }
-}
-
-.form-sign-in,
-.form-sign-up{
-  margin-bottom: 50px;
-  .ant-form-explain, .ant-form-extra {
-    position: absolute;
-  }
-  .ant-row.ant-form-item {
-    margin-bottom: 0;
-  }
-}
