--- src/components/User/components/SignInForm/index.js
@@ -1,3 +1,0 @@
-import SignInForm from './container/SignInForm';
-
-export default SignInForm;
