--- src/components/User/components/SignUpForm/index.js
@@ -1,3 +1,0 @@
-import SignUpForm from './container/SignUpForm';
-
-export default SignUpForm;
