--- src/components/User/container/User.js
@@ -1,16 +1,0 @@
-import { connect } from 'react-redux';
-import { logOut } from '@/reducers/user';
-import User from '../layout/User';
-
-const mapStateToProps = state => ({
-  ...state.user,
-});
-
-const mapDispatchToProps = {
-  logOut,
-};
-
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps,
-)(User);
