--- src/components/User/index.js
@@ -1,3 +1,0 @@
-import User from './container/User';
-
-export default User;
