+++ src/components/pages/AddNewArticle/index.js
@@ -1,0 +1,3 @@
+import AddNewArticle from './container/AddNewArticle';
+
+export default AddNewArticle;
