+++ src/components/pages/SignIn/index.js
@@ -1,0 +1,3 @@
+import LogIn from './container/SignIn';
+
+export default LogIn;
