+++ src/components/pages/SignUp/index.js
@@ -1,0 +1,3 @@
+import SignUpForm from './container/SignUp';
+
+export default SignUpForm;
