+++ src/layouts/CoreLayout/index.js
@@ -1,0 +1,3 @@
+import CoreLayout from './container/CoreLayout';
+
+export default CoreLayout;
